+++ src/blocks/text.js
@@ -40,8 +40,9 @@ const textToMjml = (item, encloseInSection) => {
 
     let mjText = `<mj-text ${properties}>${content}</mj-text>`;
 
+    // mj-text NOT allowed in mj-body
     return encloseInSection
-        ? `<mj-section><mj-column>${mjText}</mj-column></mj-section>`
+        ? `<mj-section padding="0px"><mj-column padding="0px">${mjText}</mj-column></mj-section>`
         : mjText;
 };