+++ src/blocks/vspacer.js
@@ -1,6 +1,6 @@
 import {propertiesToText, translateProps} from './common/base';
 
-const vspacerToMjml = (item) => {
+const vspacerToMjml = (item, encloseInSection) => {
     let keyTranslations = {
         'backgroundColor': 'container-background-color',
     };
@@ -9,10 +9,15 @@ const vspacerToMjml = (item) => {
 
     let properties = propertiesToText(mjmlProperties, []);
 
-    return `<mj-spacer ${properties}/>`;
+    let mjSpacer =`<mj-spacer ${properties}/>`;
+
+    // mj-spacer NOT allowed in mj-body
+    return encloseInSection
+        ? `<mj-section padding="0px"><mj-column padding="0px">${mjSpacer}</mj-column></mj-section>`
+        : mjSpacer;
 };
 
-const vspacerToText = (item) => {
+const vspacerToText = (item, isAtMainLevel) => {
     return '-----\r\n';
 };